requirements.txt
psutil
matplotlib
pillow